pyproject.toml
[project]
name = "agendador-backend"
version = "0.1.0"
description = "Backend do Agendador — SaaS de agendamento com WhatsApp"
requires-python = ">=3.12"
dependencies = [
    # ── Django ────────────────────────────────────────────────────────────────
    "django>=5.1,<6.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.6",
    "django-filter>=24.3",

    # ── Banco de dados ────────────────────────────────────────────────────────
    "psycopg[binary]>=3.2",           # PostgreSQL driver (psycopg3)

    # ── Cache / Filas ─────────────────────────────────────────────────────────
    "redis>=5.2",
    "celery[redis]>=5.4",
    "django-celery-beat>=2.7",

    # ── HTTP / Externos ───────────────────────────────────────────────────────
    "httpx>=0.28",                    # Cliente HTTP async (WhatsApp API)

    # ── Storage ───────────────────────────────────────────────────────────────
    "django-storages[s3]>=1.14",
    "Pillow>=11.0",                   # Processamento de imagens

    # ── Configuração ──────────────────────────────────────────────────────────
    "pydantic-settings>=2.7",         # Validação de variáveis de ambiente
    "python-decouple>=3.8",           # Leitura de .env

    # ── Observabilidade ───────────────────────────────────────────────────────
    "structlog>=24.4",
    "sentry-sdk[django]>=2.19",

    # ── Segurança ─────────────────────────────────────────────────────────────
    "argon2-cffi>=23.1",              # Hash de senhas (mais seguro que bcrypt)
    "cryptography>=44.0",             # Criptografia de tokens OAuth/dados sensíveis
]

[dependency-groups]
dev = [
    # ── Testes ────────────────────────────────────────────────────────────────
    "pytest>=8.3",
    "pytest-django>=4.9",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "factory-boy>=3.3",
    "faker>=33.1",
    "freezegun>=1.5",                 # Congela o tempo em testes

    # ── Qualidade de código ───────────────────────────────────────────────────
    "ruff>=0.9",
    "mypy>=1.14",
    "django-stubs[compatible-mypy]>=5.1",
    "djangorestframework-stubs>=3.15",
    "types-redis>=4.6",

    # ── Utilitários de dev ────────────────────────────────────────────────────
    "ipython>=8.30",
    "django-extensions>=3.2",
]

# ── Ruff (Linter + Formatter) ─────────────────────────────────────────────────
[tool.ruff]
target-version = "py312"
line-length = 100
src = [".", "apps", "config", "core"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "SIM", # flake8-simplify
    "RUF", # Ruff-specific rules
    "DJ",  # flake8-django
    "S",   # flake8-bandit (segurança)
    "T20", # flake8-print (evita print() em prod)
    "PIE", # flake8-pie
    "PTH", # flake8-use-pathlib
]
ignore = [
    "S101",   # allow assert em testes
    "S106",   # allow hardcoded passwords em testes
    "DJ001",  # allow nullable CharField (há casos válidos)
    "RUF012", # allow mutable class attrs no Django
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "T20"]
"**/migrations/**" = ["E501", "RUF012"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config", "core"]
known-django = ["django", "rest_framework"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

# ── MyPy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

# Permite import de módulos sem stubs
ignore_missing_imports = false
warn_unused_ignores = true
warn_redundant_casts = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "kombu.*",
    "factory.*",
    "faker.*",
    "decouple.*",
    "storages.*",
    "structlog.*",
    "argon2.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["*.migrations.*"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=apps",
    "--cov=core",
    "--cov-report=term-missing",
    "--cov-report=html:coverage_html",
    "--cov-fail-under=70",
]
markers = [
    "unit: Testes unitários (sem banco de dados)",
    "integration: Testes de integração (com banco de dados)",
    "slow: Testes lentos (ex: envio de email, WhatsApp mock)",
]

# ── Coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source = ["apps", "core"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "manage.py",
    "config/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
]
